move component dir's layout to layout base dir
